fix(format): pass varargs the types their conversions promise
Every one of these is a varargs type mismatch, which is undefined behaviour
rather than a cosmetic complaint, and the simulator build already asks for
-Wformat=2 and then discards the answer.

The ten serializer sites are one line copy-pasted per app: "%.*s" takes its
precision as int, and a size_t argument leaves printf reading the wrong width
off the stack. The width happens to fit on LP64; that is luck, not a contract.

The rest are uint32_t and int32_t printed as long. Same value on the host,
wrong type on the wire, and wrong outright wherever uint32_t is not int. The
PRIu32 macros are already how the Examples serializers do this.

JsonStreamReader indexes with size_t and printed it as %u, which truncates the
subquery past four billion elements. Unreachable in practice, still wrong.

Confirmed against g++ -Wall -Wextra -Wformat=2: the previous forms warn, these
do not. Leaves the -Wformat-truncation site in VariantConfig.cpp alone; that
one is fixed in the variant-identity change.

+++ Examples/Apps/Cycling/Software/Libs/Sources/ActivitySummarySerializer.cpp
@@ -34,7 +34,7 @@ bool ActivitySummarySerializer::save(const ActivitySummary& summary)
     const char* slash = strrchr(mPath, '/');
     if (slash) {
         char buff[SDK::Interface::IFileSystem::skMaxPathLen]{ };
-        snprintf(buff, sizeof(buff), "%.*s", static_cast<size_t>(slash - mPath), mPath);
+        snprintf(buff, sizeof(buff), "%.*s", static_cast<int>(slash - mPath), mPath);
         if (!mKernel.fs.mkdir(buff)) {
             return false;
         }
